Revert "Avinash console app"
